# API Interface/user-manager-swagger (14-11-2024).yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8093
    description: Generated server url
paths:
...
  /api/admin/roles:
    get:
      tags:
        - admin-controller
      summary: Retrieve all user roles from Keycloak
      operationId: getAllUserRoles
      responses:
        '200':
          description: User roles retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoListString'
        '400':
          description: 'Invalid request: either credentials or token must be provided!'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '401':
          description: Authentication process failed!
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoListString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/admin/pilots:
    get:
      tags:
        - admin-controller
      summary: Retrieve all pilots from Keycloak
      operationId: getAllPilots
      responses:
        '200':
          description: Pilot codes retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoListString'
        '400':
          description: 'Invalid request: either credentials or token must be provided!'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '401':
          description: Authentication process failed!
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoListString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
  /api/admin/pilot-roles:
    get:
      tags:
        - admin-controller
      summary: Retrieve all pilot roles from Keycloak
      operationId: getAllPilotRoles
      responses:
        '200':
          description: Pilot roles retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoListString'
        '400':
          description: 'Invalid request: either credentials or token must be provided!'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '401':
          description: Authentication process failed!
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoListString'
        '403':
          description: Invalid authorization parameters. Check JWT or CSRF Token
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoMapStringString'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInfoString'
...
components:
  schemas:
    ApiResponseInfoMapStringString:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            type: string
        errors:
          type: object
        message:
          type: string
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
    ApiResponseInfoString:
      type: object
      properties:
        data:
          type: string
        errors:
          type: object
        message:
          type: string
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
...
    ApiResponseInfoListString:
      type: object
      properties:
        data:
          type: array
          items:
            type: string
        errors:
          type: object
        message:
          type: string
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time